:recycle: Extract passphraseBackup component to use also for 2nd passphrase
